[FIX] _update_routes now handle multiple records (#325)
This method is called from the write() method the one how support
multiple records an api.multi method.

I was making some test when I tried to activate the mto+mts option for
multiple warehouse and the next error appears: ``ValueError: Expected singleton``

In order to fix this error I only added a loop to manage the multiple
registers.